Adjust cc violation limits
Changing the CCN and WMC calculation results in higher CCN and
especially WMC values.

Therefore the threshold for violations should to be adjusted too.
For the CCN 10 as a threshold is suggested by McCabe itself.
For the WMC no source could be found, but in other open source metric
tools 50 has been found as threshold.

Applying this thresholds on symfony leeds to the following violations:

|                            | CCN violations | WMC violations |
| -------------------------- | -------------- | -------------- |
| Before changing ccn        |       104      |       237      |
| After changing ccn         |       297      |       458      |
| After adjusting thresholds |       115      |       293      |

Altogether when applying the new CCN calculation with the new
thresholds, which are based on resources, the number of violations does
not change significantly.