Allow usage behind proxy
It reads the environment variable `http_proxy` before trying to download the composer.json dependency from `Packagist` and creates a stream context to download it if it was configured.

Typically, if you are trying to analyze your code behind a proxy, this tool is stuck getting the packagist dependency tree for all the project dependencies. With this change, its success to download them and move on with the analysis.